refactor schema settings into its own module
