+++ app/packages/core/src/components/Schema/SchemaSelection.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { Box } from "@mui/material";
+
+interface Props {
+  test?: boolean;
+}
+
+export const SchemaSelection = (props: Props) => {
+  const { test } = props;
+
+  return (
+    <Box style={{ display: "flex", position: "relative" }}>Field Selection</Box>
+  );
+};
