+++ app/packages/core/src/components/Schema/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SchemaSettings";
