+++ app/packages/core/src/components/Sidebar/Sidebar.tsx
@@ -1,18 +1,11 @@
-import React, {
-  Suspense,
-  useCallback,
-  useRef,
-  useState,
-  Fragment,
-} from "react";
+import React, { Suspense, useCallback, useRef, useState } from "react";
 import { animated, Controller, config } from "@react-spring/web";
 import styled from "styled-components";
-import { Button } from "@fiftyone/components";
 
 import { move } from "@fiftyone/utilities";
 
-import { useEventHandler, useSchemaSettings } from "@fiftyone/state";
-import { scrollbarStyles, TabOption } from "../utils";
+import { useEventHandler } from "@fiftyone/state";
+import { scrollbarStyles } from "../utils";
 import { Resizable } from "re-resizable";
 import { useRecoilState, useRecoilValue, useResetRecoilState } from "recoil";
 import { replace } from "./Entries/GroupEntries";
@@ -21,28 +14,9 @@ import * as fos from "@fiftyone/state";
 import { Box } from "@mui/material";
 import ViewSelection from "./ViewSelection";
 import { resizeHandle } from "./Sidebar.module.css";
-import CloseIcon from "@mui/icons-material/Close";
-import Checkbox from "@mui/material/Checkbox";
-import {
-  TAB_OPTIONS,
-  TAG_OPTIONS_MAP,
-} from "@fiftyone/state/src/hooks/useSchemaSettings";
+import SchemaSettings from "../Schema/SchemaSettings";
 
 const MARGIN = 3;
-
-// TODO: move
-const ModalWrapper = styled.div`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: 1000000;
-  align-items: center;
-  display: flex;
-  justify-content: center;
-  background-color: ${({ theme }) => theme.neutral.softBg};
-`;
 
 const fn = (
   items: InteractiveItems,
@@ -425,16 +399,6 @@ const Container = animated(styled.div`
   }
 `);
 
-const NewContainer = styled(Container)`
-  margin: 1rem;
-  min-width: 60vw;
-  max-width: 60vw;
-  height: 80vh;
-  overflow-y: auto;
-  min-height: auto;
-  background: white;
-`;
-
 type RenderEntry = (
   key: string,
   group: string,
@@ -474,29 +438,7 @@ const InteractiveSidebar = ({
     () => new Controller({ minHeight: 0 })
   );
 
-  const schemaModalRef = useRef<HTMLDivElement>(null);
   const modalContainer = document.getElementById("modal");
-
-  const {
-    settingModal,
-    setSettingsModal,
-    searchTerm,
-    setSearchTerm,
-    finalSelectedPaths,
-    setSelectedTab,
-    originalSelectedPaths,
-    setSelectedPaths,
-    selectedTab,
-    fieldsOnly,
-    setFieldsOnly,
-    setView,
-    toggleSelection,
-    finalSchema,
-    allFieldsChecked,
-    setAllFieldsChecked,
-  } = useSchemaSettings();
-
-  const { open: isSettingsModalOpen } = settingModal;
 
   if (entries instanceof Error) {
     throw entries;
@@ -787,256 +729,7 @@ const InteractiveSidebar = ({
         [resizableSide]: resizeHandle,
       }}
     >
-      {modalContainer && isSettingsModalOpen && (
-        <Fragment>
-          <ModalWrapper
-            ref={schemaModalRef}
-            onClick={(event) =>
-              event.target === schemaModalRef.current && console.log("TODO")
-            }
-          >
-            <NewContainer
-              style={{
-                ...screen,
-                zIndex: 10001,
-                padding: "1.5rem",
-                backgroundColor: theme.background.level2,
-              }}
-            >
-              <Box
-                style={{
-                  position: "relative",
-                  display: "flex",
-                  justifyContent: "space-between",
-                }}
-              >
-                <h3
-                  color={theme.text.primary}
-                  style={{
-                    padding: 0,
-                    margin: 0,
-                    width: "100%",
-                  }}
-                >
-                  Schema field visibilty
-                </h3>
-                <CloseIcon
-                  sx={{
-                    color: theme.text.primary,
-                    cursor: "pointer",
-                  }}
-                  onClick={() => {
-                    setSearchTerm("");
-                    setSelectedPaths(originalSelectedPaths);
-                    setSettingsModal({ ...settingModal, open: false });
-                  }}
-                />
-              </Box>
-              <Box
-                sx={{
-                  position: "relative !important",
-                  overflow: "hidden",
-                  width: "100%",
-                  paddingTop: "0.5rem",
-                }}
-              >
-                <TabOption
-                  active={selectedTab}
-                  options={TAB_OPTIONS.map((value) => {
-                    return {
-                      text: value,
-                      title: `Fiele ${value}`,
-                      onClick: () => setSelectedTab(value),
-                    };
-                  })}
-                />
-              </Box>
-              {selectedTab === TAG_OPTIONS_MAP.SEARCH && (
-                <Box style={{ display: "flex", position: "relative" }}>
-                  <input
-                    value={searchTerm}
-                    placeholder="search by fields and attributes"
-                    style={{ color: "#232323", width: "100%" }}
-                    onChange={(e) => setSearchTerm(e.target.value)}
-                  />
-                </Box>
-              )}
-              {selectedTab === TAG_OPTIONS_MAP.SELECTION && (
-                <Box
-                  display="flex"
-                  flexDirection="column"
-                  sx={{ position: "relative !important" }}
-                >
-                  <Box
-                    style={{
-                      position: "relative",
-                      padding: "0.5rem 0",
-                      color: theme.text.primary,
-                      display: "flex",
-                    }}
-                  >
-                    Show attributes
-                    <Checkbox
-                      name={"Carousel"}
-                      value={!fieldsOnly}
-                      checked={!fieldsOnly}
-                      onChange={() => setFieldsOnly(!fieldsOnly)}
-                      style={{ padding: "4px" }}
-                    />
-                  </Box>
-                  {!allFieldsChecked && (
-                    <Box
-                      style={{
-                        position: "relative",
-                        color: theme.text.primary,
-                        display: "flex",
-                      }}
-                    >
-                      Select All
-                      <Checkbox
-                        name={"Carousel"}
-                        value={allFieldsChecked}
-                        checked={allFieldsChecked}
-                        onChange={() => setAllFieldsChecked(!allFieldsChecked)}
-                        style={{ padding: "4px" }}
-                      />
-                    </Box>
-                  )}
-                  {allFieldsChecked && (
-                    <Box
-                      style={{
-                        position: "relative",
-                        color: theme.text.primary,
-                        display: "flex",
-                      }}
-                    >
-                      Deselect All
-                      <Checkbox
-                        name={"Carousel"}
-                        value={allFieldsChecked}
-                        checked={allFieldsChecked}
-                        onChange={() => setAllFieldsChecked(!allFieldsChecked)}
-                        style={{ padding: "4px" }}
-                      />
-                    </Box>
-                  )}
-                  <Box
-                    style={{
-                      position: "relative",
-                      height: "50vh",
-                      marginTop: "1rem",
-                      overflow: "auto",
-                      color: "#232323",
-                      border: `1px solid ${theme.primary.plainBorder}`,
-                    }}
-                  >
-                    {finalSchema.map((item) => {
-                      const {
-                        path,
-                        count,
-                        isSelected,
-                        pathLabelFinal,
-                        skip,
-                        disabled,
-                      } = item;
-
-                      if (skip) return null;
-
-                      return (
-                        <Box
-                          style={{
-                            padding: "0.25rem 0.25rem",
-                            borderBottom: `1px solid ${theme.primary.plainBorder}`,
-                            display: "flex",
-                          }}
-                        >
-                          <Box>
-                            <Checkbox
-                              name={"Carousel"}
-                              value={path}
-                              checked={isSelected}
-                              onChange={() => {
-                                toggleSelection(path, isSelected);
-                              }}
-                              style={{
-                                padding: 0,
-                              }}
-                              disabled={disabled}
-                            />
-                          </Box>
-                          <Box
-                            style={{
-                              paddingLeft: `${(count - 1) * 15 + 5}px`,
-                              color: disabled
-                                ? theme.text.tertiary
-                                : theme.text.primary,
-                            }}
-                          >
-                            {pathLabelFinal}
-                          </Box>
-                        </Box>
-                      );
-                    })}
-                  </Box>
-                  <Box
-                    style={{
-                      position: "relative",
-                      display: "flex",
-                      padding: "1rem 0.25rem",
-                    }}
-                  >
-                    <Button
-                      style={{
-                        color: theme.text.primary,
-                        marginRight: "0.5rem",
-                        boxShadow: "none",
-                        padding: "0.5rem 0.5rem",
-                        borderRadius: "4px",
-                      }}
-                      onClick={() => {
-                        const stageKwargs = [
-                          ["field_names", [...finalSelectedPaths]],
-                          ["_allow_missing", true],
-                        ];
-                        const stageCls = "fiftyone.core.stages.SelectFields";
-                        const stage = {
-                          _cls: stageCls,
-                          kwargs: stageKwargs,
-                        } as Stage;
-                        try {
-                          setView([stage]);
-                        } catch (e) {
-                          console.log("error", e);
-                        } finally {
-                          setSettingsModal({ open: false });
-                        }
-                      }}
-                    >
-                      Add to view
-                    </Button>
-                    <Button
-                      text="cancel"
-                      style={{
-                        color: theme.text.primary,
-                        boxShadow: "none",
-                        padding: "0.5rem 0.5rem",
-                        borderRadius: "4px",
-                      }}
-                      onClick={() => {
-                        setSettingsModal({ open: false });
-                        setSearchTerm("");
-                        setSelectedPaths(originalSelectedPaths);
-                      }}
-                    >
-                      cancel
-                    </Button>
-                  </Box>
-                </Box>
-              )}
-            </NewContainer>
-          </ModalWrapper>
-        </Fragment>
-      )}
+      {modalContainer && <SchemaSettings />}
       {!modal && (
         <Suspense>
           <Box
